refactor: only one reportsResponseToStatisticsRepo-call independent of number of dimensions tracked
