+++ naiserator-dev.json
@@ -1,3 +1,5 @@
 {
-  "ingress": ["https://pam-ad-statistics-demo.nais.oera-q.local"]
+  "ingress": [
+    "https://pam-ad-statistics-demo.nais.oera-q.local"
+  ]
 }
