+++ naiserator-prod.json
@@ -1,3 +1,5 @@
 {
-  "ingress": ["https://pam-ad-statistics-demo.nais.oera.no"]
+  "ingress": [
+    "https://pam-ad-statistics-demo.nais.oera.no"
+  ]
 }
