+++ src/main/kotlin/no/nav/arbeidsplassen/analytics/ad/AdAnalyticsController.kt
@@ -1,12 +1,12 @@
 package no.nav.arbeidsplassen.analytics.ad
 
 import no.nav.arbeidsplassen.analytics.ad.dto.AdDto
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import org.springframework.http.HttpStatus
-import org.springframework.http.ResponseEntity
 
 @RestController
 @RequestMapping("ad")
@@ -21,7 +21,6 @@ class AdAnalyticsController(
         return adAnalyticsRepository.getDtoFromUUID(UUID)
     }
 
-
     @GetMapping("/internal/isAlive")
     fun isAlive(): ResponseEntity<String> =
         ResponseEntity("OK", HttpStatus.OK)
@@ -29,5 +28,4 @@ class AdAnalyticsController(
     @GetMapping("/internal/isReady")
     fun isReady(): ResponseEntity<String> =
         ResponseEntity("OK", HttpStatus.OK)
-
 }
